Copy the changes in ALL to the local files. Put vac into vac_agent
