+++ src/examples/vac_gridworld.py
@@ -17,8 +17,7 @@ from sharedautonomy.baseline import evaluate_shared_autonomy_tabular
 
 from environments.gridworld_environment import GridworldEnvironment  # pylint: disable=import-error
 
-from all.experiments import run_experiment
-
+from utils.run_experiment import run_experiment
 
 env = GridworldEnvironment(grid_dims = (3, 3), start_state = 0, end_states = [8], obstacle_states = [], water_states = [3])
 #env = GridworldEnvironment(grid_dims = (5, 5), start_state = 0, end_states = [24], obstacle_states = [12, 17], water_states = [22])
